Modified AerobicactivityService and Add VW

+++ MountainTrainingApp.Data.Models/AerobicWorkout.cs
@@ -29,7 +29,7 @@
         public double Distance { get; set; }
 
         [Required]
-        public double BurnedCalories { get; set; }
+        public int BurnedCalories { get; set; }
 
         [Required]
         public int ElevationGain { get; set; }
